add inner product test
